% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r, R/dtensor.r, R/sptensor.r
\docType{methods}
\name{allsubs}
\alias{allsubs}
\alias{allsubs,dtensor-method}
\alias{allsubs,sptensor-method}
\title{Subscripts of all values in a tensor}
\usage{
allsubs(x)

\S4method{allsubs}{dtensor}(x)

\S4method{allsubs}{sptensor}(x)
}
\arguments{
\item{x}{tensor}
}
\description{
Subscripts of all values in a tensor
}
\keyword{internal}
