% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{array_index}
\alias{array_index}
\title{Array Index}
\usage{
array_index(x, dims)
}
\arguments{
\item{x}{vector of linear indices}

\item{dims}{dimensions}
}
\description{
Convert a vector of linear indices to a a matrix
 of array indices. Each linear index will map to
 a column in the matrix.
}
\keyword{internal}
