% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtensor-convert.r, R/sptensor-convert.r,
%   R/tensor.r
\docType{methods}
\name{as.vector,dtensor-method}
\alias{as.vector,dtensor-method}
\alias{as.vector,sptensor-method}
\alias{as.vector}
\title{Vectorize values of a tensor}
\usage{
\S4method{as.vector}{dtensor}(x)

\S4method{as.vector}{sptensor}(x)
}
\arguments{
\item{x}{tensor}
}
\description{
Vectorize values of a tensor
}
