% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r, R/dtensor-convert.r
\docType{methods}
\name{as_dtensor}
\alias{as_dtensor}
\alias{as_dtensor,sptensor-method}
\alias{as_dtensor,array-method}
\title{Convert objects to sparse or dense tensors}
\usage{
as_dtensor(x, ...)

\S4method{as_dtensor}{sptensor}(x)

\S4method{as_dtensor}{array}(x)
}
\arguments{
\item{x}{object}

\item{...}{extra params}
}
\description{
Convert objects to sparse or dense tensors
}
