% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r, R/sptensor-convert.r
\docType{methods}
\name{as_sptensor}
\alias{as_sptensor}
\alias{as_sptensor,sptensor-method}
\alias{as_sptensor,dtensor-method}
\alias{as_sptensor,data.frame-method}
\title{Convert objects to sparse tensors}
\usage{
as_sptensor(x, ...)

\S4method{as_sptensor}{sptensor}(x)

\S4method{as_sptensor}{dtensor}(x)

\S4method{as_sptensor}{data.frame}(x, valcol = NULL, dims = NULL)
}
\arguments{
\item{x}{object}

\item{...}{extra params}

\item{valcol}{column to use for the tensor values. all other columns are treated as indices}

\item{dims}{dimensions of tensor. If not provided, the maximum value for each of the indices is used.}
}
\description{
Convert objects to sparse tensors
}
