% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asserts.r
\name{assert_subs_within_dims}
\alias{assert_subs_within_dims}
\title{Are subscripts within dimensions?}
\usage{
assert_subs_within_dims(subs, dims)
}
\arguments{
\item{subs}{subscripts}

\item{dims}{dimensions}
}
\description{
Checks that all subscripts have coordinates within the dimension sizes
}
\keyword{internal}
