% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{build_indices}
\alias{build_indices}
\title{convert args i,j,... to array of coords.}
\usage{
build_indices(dims, ...)
}
\arguments{
\item{dims}{tensor dimensions}

\item{...}{indices}
}
\description{
substitute missing args k with 1:dim[k]
}
\keyword{internal}
