% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtensor.r, R/sptensor.r, R/tensor.r
\docType{methods}
\name{dim,dtensor-method}
\alias{dim}
\alias{dim,dtensor-method}
\alias{dim,sptensor-method}
\title{Dimensions of a tensor}
\usage{
\S4method{dim}{dtensor}(x)

\S4method{dim}{sptensor}(x)
}
\arguments{
\item{x}{tensor}
}
\description{
Dimensions of a tensor
}

