% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtensor.r, R/sptensor.r, R/tensor.r
\docType{methods}
\name{dimnames,dtensor-method}
\alias{dimnames,dtensor-method}
\alias{dimnames<-,dtensor,list-method}
\alias{dimnames<-,dtensor-method}
\alias{dimnames<-,dtensor,NULL-method}
\alias{dimnames<-,dtensor-method,NULL}
\alias{dimnames<-,dtensor,ANY-method}
\alias{dimnames<-,dtensor-method,ANY}
\alias{dimnames,sptensor-method}
\alias{dimnames<-,sptensor,list-method}
\alias{dimnames<-,sptensor-method,list}
\alias{dimnames<-,sptensor,NULL-method}
\alias{dimnames<-,sptensor-method,NULL}
\alias{dimnames<-,sptensor,ANY-method}
\alias{dimnames<-,sptensor-method,ANY}
\alias{dimnames}
\title{Dimension names of a tensor}
\usage{
\S4method{dimnames}{dtensor}(x)

\S4method{dimnames}{dtensor,list}(x) <- value

\S4method{dimnames}{dtensor,`NULL`}(x) <- value

\S4method{dimnames}{dtensor,ANY}(x) <- value

\S4method{dimnames}{sptensor}(x)

\S4method{dimnames}{sptensor,list}(x) <- value

\S4method{dimnames}{sptensor,`NULL`}(x) <- value

\S4method{dimnames}{sptensor,ANY}(x) <- value
}
\arguments{
\item{x}{tensor}

\item{value}{replacement value. Must be a list of length equal to the number of dimensions in the tensor}
}
\description{
Dimension names of a tensor
}
