% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{dimnames_to_indices}
\alias{dimnames_to_indices}
\title{Convert dimnames to tensor indices}
\usage{
dimnames_to_indices(x, dnames, dimensions)
}
\arguments{
\item{x}{tensor}

\item{dnames}{selected dimnames}

\item{dimensions}{selected dimensions}
}
\description{
Convert dimnames to tensor indices
}
\keyword{internal}
