% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{dots_arg}
\alias{dots_arg}
\title{Get specific value from ellipsis arguments}
\usage{
dots_arg(argname, default, ...)
}
\arguments{
\item{argname}{name of argument}

\item{default}{value to return if argname not in ...}

\item{...}{ellipsis arguments}
}
\description{
Looks for \code{argname} in ellipsis args. If it is
not there, returns \code{default}.
}
