% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.r
\docType{class}
\name{dtensor-class}
\alias{dtensor-class}
\title{An S4 class for a dense tensor}
\description{
Simple wrapper around an n-dimensional array.
}
\section{Slots}{

\describe{
\item{\code{x}}{n dimensional array}
}}

