% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtensor-group-generics.r
\docType{methods}
\name{dtensor-Complex}
\alias{dtensor-Complex}
\alias{Complex,dtensor-method}
\title{Complex Methods for dense tensors}
\usage{
\S4method{Complex}{dtensor}(z)
}
\arguments{
\item{z}{dense tensor}
}
\description{
Complex Methods for dense tensors
}
\keyword{internal}
