% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtensor-extract.r
\docType{methods}
\name{dtensor-extract}
\alias{dtensor-extract}
\alias{[,dtensor,missing,missing,ANY-method}
\alias{[,dtensor,missing,missing-method}
\alias{[,dtensor,numeric,missing,ANY-method}
\alias{[,dtensor,numeric,missing-method}
\alias{[,dtensor,missing,numeric,ANY-method}
\alias{[,dtensor,missing,numeric-method}
\alias{[,dtensor,numeric,numeric,ANY-method}
\alias{[,dtensor,numeric,numeric-method}
\alias{[,dtensor,list,missing,ANY-method}
\alias{[,dtensor,list,missing-method}
\alias{[,dtensor,matrix,missing,ANY-method}
\alias{[,dtensor,matrix,missing-method}
\title{Extract values from a dense tensor}
\usage{
\S4method{[}{dtensor,missing,missing,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{dtensor,numeric,missing,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{dtensor,missing,numeric,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{dtensor,numeric,numeric,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{dtensor,list,missing,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{dtensor,matrix,missing,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{dtensor object}

\item{i}{numeric index, vector, list, or matrix}

\item{j}{numeric index}

\item{...}{additional numeric indices}

\item{drop}{whether to drop dimensions}
}
\description{
Extract values from a dense tensor
}
