% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtensor-group-generics.r
\docType{methods}
\name{dtensor-Math}
\alias{dtensor-Math}
\alias{Math,dtensor-method}
\title{Math Methods for dense tensors}
\usage{
\S4method{Math}{dtensor}(x)
}
\arguments{
\item{x}{dense tensor}
}
\description{
Math Methods for dense tensors
}
\keyword{internal}
