% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtensor-ops.r
\docType{methods}
\name{dtensor-Math2}
\alias{dtensor-Math2}
\alias{Math2,dtensor-method}
\title{Math2 Methods for dense tensors}
\usage{
\S4method{Math2}{dtensor}(x, digits)
}
\arguments{
\item{x}{dense tensor}

\item{digits}{number of digits to be used in round or signif.}
}
\description{
Math2 Methods for dense tensors
}
