% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtensor-group-generics.r
\docType{methods}
\name{dtensor-Ops}
\alias{dtensor-Ops}
\alias{Ops,dtensor,dtensor-method}
\alias{Ops,dtensor,sptensor-method}
\alias{Ops,sptensor,dtensor-method}
\alias{Ops,dtensor,missing-method}
\alias{Ops,dtensor,ANY-method}
\alias{Ops,ANY,dtensor-method}
\title{Ops Methods for dense tensors}
\usage{
\S4method{Ops}{dtensor,dtensor}(e1, e2)

\S4method{Ops}{dtensor,sptensor}(e1, e2)

\S4method{Ops}{sptensor,dtensor}(e1, e2)

\S4method{Ops}{dtensor,missing}(e1, e2)

\S4method{Ops}{dtensor,ANY}(e1, e2)

\S4method{Ops}{ANY,dtensor}(e1, e2)
}
\arguments{
\item{e1, e2}{at least one of them a dense tensor}
}
\description{
Ops Methods for dense tensors
}
\keyword{internal}
