% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtensor-replace.r
\docType{methods}
\name{dtensor-replace}
\alias{dtensor-replace}
\alias{[<-,dtensor,ANY,ANY,ANY-method}
\alias{[<-,dtensor,ANY,ANY-method}
\alias{[<-,dtensor,ANY,missing,ANY-method}
\alias{[<-,dtensor,ANY,missing-method}
\alias{[<-,dtensor,matrix,missing,ANY-method}
\alias{[<-,dtensor,matrix,missing-method}
\alias{[<-,dtensor,list,missing,ANY-method}
\alias{[<-,dtensor,list,missing-method}
\title{Replace values from a dense tensor}
\usage{
\S4method{[}{dtensor,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{dtensor,ANY,missing,ANY}(x, i, j, ...) <- value

\S4method{[}{dtensor,matrix,missing,ANY}(x, i, j, ...) <- value

\S4method{[}{dtensor,list,missing,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{dtensor object}

\item{i}{numeric index, vector, list, or matrix}

\item{j}{numeric index}

\item{...}{additional numeric indices}

\item{value}{replacement value(s)}
}
\description{
Replace values from a dense tensor
}
