% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtensor-group-generics.r
\docType{methods}
\name{dtensor-Summary}
\alias{dtensor-Summary}
\alias{Summary,dtensor-method}
\title{Summary Methods for dense tensors}
\usage{
\S4method{Summary}{dtensor}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{dense tensor}

\item{...}{further arguments passed to or from methods.}

\item{na.rm}{logical: should missing values be removed?}
}
\description{
Summary Methods for dense tensors
}
\keyword{internal}
