% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r, R/dtensor.r
\docType{methods}
\name{dtensor}
\alias{dtensor}
\alias{dtensor,array-method}
\alias{dtensor,numeric-method}
\title{Construct a dense tensor}
\usage{
dtensor(x)

\S4method{dtensor}{array}(x)

\S4method{dtensor}{numeric}(x)
}
\arguments{
\item{x}{n-dimensional R array}
}
\description{
Construct a \linkS4class{dtensor} from a multi-dimensional array.
}
\examples{
# A 2x2x2 dense tensor
arr <- array(data = c(1,0,0,0,1,0,0,0) , dim = c(2,2,2))
X <- dtensor(arr)

}
\seealso{
\code{\link{dtensor-class}} for class documentation.
}

