% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptensor-extract.r
\name{extract_vec}
\alias{extract_vec}
\title{Extract indices from a sparse tensor}
\usage{
extract_vec(x, idxmat)
}
\arguments{
\item{x}{sptensor}

\item{idxmat}{matrix of indices}
}
\description{
Extracts indices from tensor and returns as a vector.
}
\keyword{internal}
