% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{fill_missing_indices}
\alias{fill_missing_indices}
\title{Fill NULL indices with a range from 1:dim}
\usage{
fill_missing_indices(index, dim)
}
\arguments{
\item{index}{a numeric value or NULL}

\item{dim}{size of dimension}
}
\description{
Fill NULL indices with a range from 1:dim
}
