% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r, R/dtensor-math.r,
%   R/sptensor-math.r
\docType{methods}
\name{innerprod}
\alias{innerprod}
\alias{innerprod,dtensor,dtensor-method}
\alias{innerprod,sptensor,sptensor-method}
\title{Calculate the inner product of a pair of tensors}
\usage{
innerprod(x, y)

\S4method{innerprod}{dtensor,dtensor}(x, y)

\S4method{innerprod}{sptensor,sptensor}(x, y)
}
\arguments{
\item{x, y}{tensors}
}
\description{
Calculate the inner product of a pair of tensors
}
