% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtensor.r, R/sptensor.r, R/tensor.r
\name{is_dtensor}
\alias{is_dtensor}
\alias{is_sptensor}
\alias{is_tensor}
\title{Test if the object is a tensor}
\usage{
is_dtensor(x)

is_sptensor(x)

is_tensor(x)
}
\arguments{
\item{x}{object}
}
\description{
Test if the object is a tensor
}
\section{Functions}{
\itemize{
\item \code{is_dtensor}: dense tensor

\item \code{is_sptensor}: sparse tensor
}}

