% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.r
\docType{methods}
\name{length,tensor-method}
\alias{length,tensor-method}
\title{Length of a tensor}
\usage{
\S4method{length}{tensor}(x)
}
\arguments{
\item{x}{tensor}
}
\description{
The total number of values in a tensor. Equal to the product of
the dimensions.
}

