% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{mat_to_listidx}
\alias{mat_to_listidx}
\title{convert matrix of indices to list of indices}
\usage{
mat_to_listidx(m)
}
\arguments{
\item{m}{matrix}
}
\description{
convert matrix of indices to list of indices
}
