% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptensor-utils.r
\name{matches}
\alias{matches}
\title{Match index in sparse tensor}
\usage{
matches(idx, x)
}
\arguments{
\item{idx}{index}

\item{x}{sptensor}
}
\description{
Matches a given index to the index of a non-zero value
in a sparse tensor.  Returns the index if there is match.
Returns 0 if there is no match. Returns NA if the index
is out of bounds.
}
\keyword{internal}
