% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r, R/dtensor-math.r,
%   R/sptensor-math.r
\docType{methods}
\name{norm}
\alias{norm}
\alias{norm,dtensor-method}
\alias{norm,sptensor-method}
\title{Calculate the Frobenius norm of a tensor}
\usage{
norm(x)

\S4method{norm}{dtensor}(x)

\S4method{norm}{sptensor}(x)
}
\arguments{
\item{x}{tensor}
}
\description{
Calculate the Frobenius norm of a tensor
}
