% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r, R/dtensor.r, R/sptensor.r
\docType{methods}
\name{nzsubs}
\alias{nzsubs}
\alias{nzsubs,dtensor-method}
\alias{nzsubs,sptensor-method}
\title{Subscripts of non-zero values in a tensor}
\usage{
nzsubs(x)

\S4method{nzsubs}{dtensor}(x)

\S4method{nzsubs}{sptensor}(x)
}
\arguments{
\item{x}{tensor}
}
\description{
Subscripts of non-zero values in a tensor
}
