% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r, R/dtensor.r, R/sptensor.r
\docType{methods}
\name{nzvals}
\alias{nzvals}
\alias{nzvals,dtensor-method}
\alias{nzvals,sptensor-method}
\title{Non-zero values in a tensor}
\usage{
nzvals(x)

\S4method{nzvals}{dtensor}(x)

\S4method{nzvals}{sptensor}(x)
}
\arguments{
\item{x}{tensor}
}
\description{
Non-zero values in a tensor
}
