% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r, R/dtensor-math.r,
%   R/sptensor-math.r
\docType{methods}
\name{outerprod}
\alias{outerprod}
\alias{outerprod,dtensor,dtensor-method}
\alias{outerprod,sptensor,sptensor-method}
\title{Calculate the outer product of a pair of tensors}
\usage{
outerprod(x, y)

\S4method{outerprod}{dtensor,dtensor}(x, y)

\S4method{outerprod}{sptensor,sptensor}(x, y)
}
\arguments{
\item{x, y}{tensors}
}
\description{
Calculate the outer product of a pair of tensors
}
