% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r, R/dtensor-unfold.r,
%   R/sptensor-unfold.r
\docType{methods}
\name{refold}
\alias{refold}
\alias{refold,unfolded_dtensor-method}
\alias{refold,unfolded_sptensor-method}
\title{Refold an unfolded tensor}
\usage{
refold(x)

\S4method{refold}{unfolded_dtensor}(x)

\S4method{refold}{unfolded_sptensor}(x)
}
\arguments{
\item{x}{an unfolded tensor}
}
\description{
Refold an unfolded tensor
}
