% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptensor-replace.r
\name{replace_sptensor}
\alias{replace_sptensor}
\title{Replace values in a sparse tensor}
\usage{
replace_sptensor(x, idxmat, value)
}
\arguments{
\item{x}{sptensor}

\item{idxmat}{matrix of indices}

\item{value}{replacement value(s)}
}
\description{
Replaces values in sparse tensor and returns a new sparse tensor.
}
