% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{row_apply}
\alias{row_apply}
\title{Apply a function to rows of a matrix}
\usage{
row_apply(x, f, ...)
}
\arguments{
\item{x}{matrix}

\item{f}{function to apply to each row}

\item{...}{extra args to f}
}
\description{
Convenience wrapper around apply. Always returns a matrix.
}
