% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptensor-utils.r
\name{setops_subs}
\alias{setops_subs}
\alias{union_subs}
\alias{intersect_subs}
\title{Apply set function to subscript arrays}
\usage{
setops_subs(subs1, subs2, f)

union_subs(subs1, subs2)

intersect_subs(subs1, subs2)
}
\arguments{
\item{subs1, subs2}{subscript arrays}

\item{f}{set function to apply. Ex. union, setdiff, intersect}
}
\description{
Apply set function to subscript arrays
}
