% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtensor.r, R/sptensor.r, R/tensor.r
\docType{methods}
\name{show,dtensor-method}
\alias{show,dtensor-method}
\alias{show,sptensor-method}
\alias{show}
\title{Show a tensor}
\usage{
\S4method{show}{dtensor}(object)

\S4method{show}{sptensor}(object)
}
\arguments{
\item{object}{tensor}
}
\description{
Show a tensor
}
