% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.r
\docType{class}
\name{sptensor-class}
\alias{sptensor-class}
\title{An S4 class for a sparse tensor}
\description{
Stores the tensor in co-ordinate (COO) format. Non-zero entries are stored
by their subscripts (i1,i2,i3,...,in) the \code{subs} matrix and their vals
in the \code{vals} vector.
}
\section{Slots}{

\describe{
\item{\code{subs}}{matrix with length(dims) rows and length(vals) cols.}

\item{\code{vals}}{values of non-zero entries.}

\item{\code{dims}}{sizes of each dimension}
}}

