% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptensor-group-generics.r
\docType{methods}
\name{sptensor-Complex}
\alias{sptensor-Complex}
\alias{Complex,sptensor-method}
\title{Complex Methods for sparse tensors}
\usage{
\S4method{Complex}{sptensor}(z)
}
\arguments{
\item{z}{sparse tensor}
}
\description{
Complex Methods for sparse tensors
}
\keyword{internal}
