% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptensor-group-generics.r
\docType{methods}
\name{sptensor-Math}
\alias{sptensor-Math}
\alias{Math,sptensor-method}
\alias{cummax,sptensor-method}
\alias{cummin,sptensor-method}
\alias{cumprod,sptensor-method}
\alias{cumsum,sptensor-method}
\title{Math Methods for sparse tensors}
\usage{
\S4method{Math}{sptensor}(x)

\S4method{cummax}{sptensor}(x)

\S4method{cummin}{sptensor}(x)

\S4method{cumprod}{sptensor}(x)

\S4method{cumsum}{sptensor}(x)
}
\arguments{
\item{x}{sparse tensor}
}
\description{
Math Methods for sparse tensors
}
\keyword{internal}
