% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptensor-group-generics.r
\docType{methods}
\name{sptensor-Math2}
\alias{sptensor-Math2}
\alias{Math2,sptensor-method}
\title{Math2 Methods for sparse tensors}
\usage{
\S4method{Math2}{sptensor}(x, digits)
}
\arguments{
\item{x}{sparse tensor}

\item{digits}{number of digits to be used in round or signif.}
}
\description{
Math2 Methods for sparse tensors
}
\keyword{internal}
