% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptensor-group-generics.r
\docType{methods}
\name{sptensor-Ops}
\alias{sptensor-Ops}
\alias{Ops,sptensor,sptensor-method}
\alias{Ops,sptensor,missing-method}
\alias{Ops,sptensor,ANY-method}
\alias{Ops,ANY,sptensor-method}
\title{Ops Methods for dense tensors}
\usage{
\S4method{Ops}{sptensor,sptensor}(e1, e2)

\S4method{Ops}{sptensor,missing}(e1, e2)

\S4method{Ops}{sptensor,ANY}(e1, e2)

\S4method{Ops}{ANY,sptensor}(e1, e2)
}
\arguments{
\item{e1, e2}{at least one of them a sparse tensor}
}
\description{
Ops Methods for dense tensors
}
\keyword{internal}
