% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptensor-replace.r
\docType{methods}
\name{sptensor-replace}
\alias{sptensor-replace}
\alias{[<-,sptensor,missing,missing,ANY-method}
\alias{[<-,sptensor,missing,missing-method}
\alias{[<-,sptensor,numeric,missing,ANY-method}
\alias{[<-,sptensor,numeric,missing-method}
\alias{[<-,sptensor,missing,numeric,ANY-method}
\alias{[<-,sptensor,missing,numeric-method}
\alias{[<-,sptensor,numeric,numeric,ANY-method}
\alias{[<-,sptensor,numeric,numeric-method}
\alias{[<-,sptensor,list,missing,ANY-method}
\alias{[<-,sptensor,list,missing-method}
\alias{[<-,sptensor,matrix,missing,ANY-method}
\alias{[<-,sptensor,matrix,missing-method}
\title{Replace values from a sparse tensor}
\usage{
\S4method{[}{sptensor,missing,missing,ANY}(x, i, j, ...) <- value

\S4method{[}{sptensor,numeric,missing,ANY}(x, i, j, ...) <- value

\S4method{[}{sptensor,missing,numeric,ANY}(x, i, j, ...) <- value

\S4method{[}{sptensor,numeric,numeric,ANY}(x, i, j, ...) <- value

\S4method{[}{sptensor,list,missing,ANY}(x, i, j, ...) <- value

\S4method{[}{sptensor,matrix,missing,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{sptensor object}

\item{i}{numeric index, vector, list, or matrix}

\item{j}{numeric index}

\item{...}{additional numeric indices}

\item{value}{replacement value(s)}
}
\description{
Replace values from a sparse tensor
}
