% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptensor-group-generics.r
\docType{methods}
\name{sptensor-Summary}
\alias{sptensor-Summary}
\alias{Summary,sptensor-method}
\title{Summary Methods for sparse tensors}
\usage{
\S4method{Summary}{sptensor}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{sparse tensor}

\item{...}{further arguments passed to or from methods.}

\item{na.rm}{logical: should missing values be removed?}
}
\description{
Summary Methods for sparse tensors
}
