% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.r
\docType{class}
\name{tensor-class}
\alias{tensor-class}
\title{Virtual S4 class for all tensor objects}
\description{
The \code{tensor} class is a base virtual tensor class for objects in this
repository.
}

