% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r, R/dtensor-unfold.r,
%   R/sptensor-unfold.r
\docType{methods}
\name{unfold}
\alias{unfold}
\alias{unfold,dtensor,numeric-method}
\alias{unfold,sptensor,numeric-method}
\title{Unfold (matricize) a tensor along a mode}
\usage{
unfold(x, mode)

\S4method{unfold}{dtensor,numeric}(x, mode)

\S4method{unfold}{sptensor,numeric}(x, mode)
}
\arguments{
\item{x}{tensor}

\item{mode}{dimension to unfold along}
}
\description{
Unfold (matricize) a tensor along a mode
}
