% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.r
\docType{class}
\name{unfolded_dtensor-class}
\alias{unfolded_dtensor-class}
\title{An S4 class for an unfolded dense tensor}
\description{
An unfolded ("matriczied") tensor along a specified dimension. Can be easily
refolded back into a tensor with command \code{refold}
}
\section{Slots}{

\describe{
\item{\code{mat}}{sparse matrix representing unfolded tensor}

\item{\code{mode}}{dimension along which tensor was unfolded}

\item{\code{tensor_dims}}{dimensions of original tensor}
}}

