% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r, R/dtensor-unfold.r
\docType{methods}
\name{unfolded_dtensor}
\alias{unfolded_dtensor}
\alias{unfolded_dtensor,Matrix,numeric,numeric-method}
\title{Construct an unfolded dense tensor}
\usage{
unfolded_dtensor(mat, mode, tensor_dims)

\S4method{unfolded_dtensor}{Matrix,numeric,numeric}(mat, mode, tensor_dims)
}
\arguments{
\item{mat}{dense matrix representing unfolded tensor}

\item{mode}{dimension to unfold tensor along}

\item{tensor_dims}{original dimensions of tensor (useful for re-folding)}
}
\description{
Construct an \linkS4class{unfolded_dtensor} from a dense matrix,
mode, and dimensions of orignal tensor.
}
\seealso{
\code{\link{unfolded_dtensor-class}} for class documentation.
}
