% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r, R/sptensor-unfold.r
\docType{methods}
\name{unfolded_sptensor}
\alias{unfolded_sptensor}
\alias{unfolded_sptensor,sparseMatrix,numeric,numeric-method}
\title{Construct an unfolded sparse tensor}
\usage{
unfolded_sptensor(mat, mode, tensor_dims)

\S4method{unfolded_sptensor}{sparseMatrix,numeric,numeric}(mat, mode,
  tensor_dims)
}
\arguments{
\item{mat}{sparse matrix representing unfolded tensor}

\item{mode}{dimension to unfold tensor along}

\item{tensor_dims}{original dimensions of tensor (useful for re-folding)}
}
\description{
Construct an \linkS4class{unfolded_sptensor} from a sparse matrix,
mode, and dimensions of orignal tensor.
}
\seealso{
\code{\link{unfolded_sptensor-class}} for class documentation.
}
