% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.r
\docType{class}
\name{unfolded_tensor-class}
\alias{unfolded_tensor-class}
\title{Virtual S4 class for unfolded tensor objects}
\description{
The \code{unfolded_tensor} class is a base virtual class for unfolded tensors in this
repository.
}
\keyword{internal}
