% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{vec_index}
\alias{vec_index}
\alias{vec_index_one}
\title{Vector Index}
\usage{
vec_index(x, dims)

vec_index_one(x, cumdims)
}
\arguments{
\item{x}{vector, matrix, or list of numeric indices}

\item{dims}{dimensions}

\item{cumdims}{cumulative size of tensor}
}
\description{
Coverts a vector, matrix, or list of array indices to
linear indices.
}
