% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r, R/dtensor.r, R/sptensor.r
\docType{methods}
\name{zsubs}
\alias{zsubs}
\alias{zsubs,dtensor-method}
\alias{zsubs,sptensor-method}
\title{Subscripts of zero values in a tensor}
\usage{
zsubs(x)

\S4method{zsubs}{dtensor}(x)

\S4method{zsubs}{sptensor}(x)
}
\arguments{
\item{x}{tensor}
}
\description{
Subscripts of zero values in a tensor
}
\keyword{internal}
