% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptensor-extract.r
\name{extract_sptensor}
\alias{extract_sptensor}
\title{Extract indices from a sparse tensor}
\usage{
extract_sptensor(x, idxmat, drop = FALSE)
}
\arguments{
\item{x}{sptensor}

\item{idxmat}{matrix of indices requested}

\item{drop}{drop dimensions of size 1?}
}
\description{
Extracts indices from a sparse tensor and returns as a new tensor.
Can also drop dimensions of the tensor that have size 1.
}
\keyword{internal}
