% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtensor-unfold.r, R/dtensor.r,
%   R/sptensor-unfold.r, R/sptensor.r, R/tensor.r
\docType{methods}
\name{show,unfolded_dtensor-method}
\alias{show,unfolded_dtensor-method}
\alias{show,dtensor-method}
\alias{show,unfolded_sptensor-method}
\alias{show,sptensor-method}
\alias{show}
\title{Show a tensor}
\usage{
\S4method{show}{unfolded_dtensor}(object)

\S4method{show}{dtensor}(object)

\S4method{show}{unfolded_sptensor}(object)

\S4method{show}{sptensor}(object)
}
\arguments{
\item{object}{tensor}
}
\description{
Show a tensor
}
\keyword{internal}
